backend/requirements.txt
fastapi
pydantic
transformers
pytesseract
pillow
uvicorn
tensorflow